rev_g: [WIP] correct principle width to fit case
Accidentally measured the distance between the *tips* of the PCB standoff
and the case screw shaft, which added 3mm in Z. The standoff and screw should
only be 8mm apart, not 11mm. With the extra 3mm, the corners of the board
were clipping into the rounded corners of the case.