rev_g: [WIP] routing; fix power switch point
Don't want to be in a situation where it's possible for the battery
to not charge becuase the power switch is off. Instead of disconnecting
the battery, disconnect the `VBAT` net from the regulator. Now if the
switch is off, USB power won't work either, but that's an obvious fix
that only happens in front of your face.