rev_h: tie !CHG LED to +3v3
This means the LED won't power on to indicate charging if the device
is switched off, but that's an uncommon case anyway. We may even end
up just DNP'ing the CHG LED; it's of middling use; in which case an
internal nRF pull-up can be used in place of this for the !CHG net.